+++ service/secretsmanager/errors.go
@@ -53,6 +53,10 @@ const (
 	//    * You tried to enable rotation on a secret that doesn't already have a
 	//    Lambda function ARN configured and you didn't include such an ARN as a
 	//    parameter in this call.
+	//
+	//    * The secret is managed by another service, and you must use that service
+	//    to update it. For more information, see Secrets managed by other Amazon
+	//    Web Services services (https://docs.aws.amazon.com/secretsmanager/latest/userguide/service-linked-secrets.html).
 	ErrCodeInvalidRequestException = "InvalidRequestException"
 
 	// ErrCodeLimitExceededException for service response error code